Flesh out first pass at ksmin.py and update Makefile to support
This commit adds a KerboScript minification tool to save disk space on my future
ship designs.

---

It has always bothered me that kOS measures disk space usage by doing a pure
character count. That's not really how the vast majority of languages in the
world work (I'm not aware of any?), but more importantly this also discourages
two things which are very important for quality software engineering:

(1): Discourages detailed comments (or indeed, any comments at all)
(2): Discourages descriptive variable and function names

Normally, I would say that allowing for (2) somewhat mitigates the need for (1),
but in this case it is easier to perform minification by stripping comments and
doing other well defined syntax transformations. Minification that requires
renaming variables/functions is a relatively complex problem and not something I
think is worth really diving into at the moment.

So with all of that being considered, I really want to focus on (1) writing well
commented code (2) using short variable/function names when possible. This will
ensure that my code minifies nicely without too much effort on my part to
improve the minifier.

Though if I do come across any small wins with the minifier I will make and test
those changes as I see fit.

---

You may be wondering, why not just compile the .ks files into .ksm and I have
two responses to that. (1) I am doing that and (2) that has *a lot* of overhead.
In some cases the .ksm files I am ending up with are actually larger than the
.ks files that I started with, and at least in what I have seen so far the .ksm
files are always larger than my minified files as well.

---

As of this commit, only the minify-all-safe option is safe to use. With the more
aggressive minify-all target some scripts work (those that do not declare
variables or functions), but others do not.

The difference between the safe and non-safe options w.r.t. file size is not
significant enough for the time I need to put into closing the gap, so this is
ok for now.