scripts: Remove compile statements from all of my scripts
After running my .ks source files through the minifier, I am not seeing *any*
benefit from running the compile command. I ran a compile on all of my lib/
scripts and for every single case the .ksm file was larger than the minified .ks
file.

This also gets rid of another limiting factor which is there there was some
amount of 'slop' in the compile process I was using, which was to copy the file
to the craft, compile it, then delete the .ks. That meant that I actually needed
2x the disk space to load a given file!

Of course that could have been (and probably should have been) mitigated by
compiling to .ksm on the KSC server and then copying the .ksm file to the craft
directly. If I actually saw any benefit from compilation, that is the process I
would have switched to.