ksx-lang: Begin supporting KerboScript Extended
I have added a keyword `@ksx` which preceeds all ksx extension keywords. At the
moment, ksx.py is just removing all `@ksx` lines but I intend to support two new
extensions soon:

* Import Functionality

I want the transpiler to be able to inline entire .ks files as well as
individual functions from those files. The syntax I want for this is below:

```
@ksx import ("example" "another_example").
@ksx from ("example") import (function_to_import).
```

The first example would find "example.ks" and "another_example.ks" files in the
search path and inline them directly on this line.

The second example would find "example.ks" on the search path and inline the
"function_to_import" function definition on this line.

In the cases of the parentheticals, these can be treated like lisp lists, the
module name (e.g. "example", or "another_example") should be in a string while
function imports should not be stringified.

* Runtime Checks

I have not written *that* many scripts so far, but I am already finding it
difficult to track the fact that some of them are written under the assumption
that another KerboScript file has previously been executed on the running CPU.

For now, I would just like a formal way to document this assumption, so am
adding the `@ksx` extension `executed` to indicate that this script expects the
given script(s) to have been executed (run, runpath) previously.

For now this is just a no-op and will transpile to a removed line, at some point
it may make sense to transpile this into a lightweight runtime check/error.
Desired syntax for this extension is below:

```
@ksx executed ("example").
```