scripts: Simplify bootstrap and beacon (esp. ship identification)
* Bootstrap UUID Simplification

Previously the bootstrap file (executed by all bootloaders) would generate a
uuid, guid, and store the ship:name in a json file on the vessel as well as
transferring that information back to the KSC. This took up precious disk space
that I wanted to reclaim so I:

0. Got rid of the identity.json file altogether by...
1. Change the uuid to be stored in the core:tag of the active kOSProcessor
2. Got rid of the guid, that's not something that I will actually need as I can
   just update my build script to generate multiple uuid targeted update files
   if needed.
3. Just use ship:name instead of storing it to the json file

This reclaims ~170B of disk space just from not storing the identity.json file
on the vessel disk, it also removes a good chunk of code from the bootloader so
saves a little bit of space there.

* Bootstrap use @ksx Features

The bootstrap script was manually loading:

- 'lib/fileop' - Because this is used by the bootloaders themselves
- 'lib/ui' - Because the notify function is used by almost every script

It was storing these files on disk the whole time. I considered updating the
bootloader to remove these two library files after executing them, but I want
the CPU to be able to cleanly restart even if it doesn't have a connection to
KSC.

I've updated bootstrap to `@ksx import` these two files instead so that they
will be inlide in bootstrap itself. This ends up saving ~140B just due to the
code savings you get from not needing to dynamically ensure these files are
downloaded.