ksx: split compiliation function so that one can work on file_lines directly
To improve unit testing, I want a compile function I can use that does not need
to load the file from disk or write the file back to disk. Instead it will take
a list of lines (as though they had been read by readlines()) and will return a
string (as the string would be written back out to a file).