fix: @ksx directives are always inserted at the top of the file
Directives such as:

```
@ksx import ("lib/telemetry").
```

Should be compiled so that the line is removed and replaced with the relevant
file or function contents inside. This was already working but the fiel contents
were being inserted at the top of the compiled file, rather than on the line
with the @ksx statement.

The same goes for:

```
@ksx from ("lib/telemetry") import (vessel_data).
```

Which should inline the vessel_data function from lib/telemetry on the current
line, but was putting it at the top of the file.

This commit fixes that behavior for both statement types.

This commit also adds tests! For now I'm just going to write tests for github
issues until I build out the test system a bit better, but it's a reasonable
starting point.

---

Fixes: #4