ksx: Partially implement recursive descent compilation
This commit implements the relatively easier part of the recursive descent
compiler which is to just continually compile the given lines until there are no
more ksx statements.

This does choke however when two files create a circular reference. I have
added a recursion limit of 6 (which may actually be too low) to catch this
possible condition. Python actually chokes at some point too, but it takes a
while and I'd rather catch it sooner (also prefer giving a more friendly error
message).

I attempted to use a hash of the file contents to track whether we have already
compiled that file, but that is not working because the thing being hashed is
accumulating each compilation step, so it changes.

Anyways, recursion is hard, and this still needs some work, but it's usable
enough as long as you don't have any circular references.

Related to issue #7